% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_c_index.R
\name{calc_c_index}
\alias{calc_c_index}
\title{Calculate C-index for the specific variables}
\usage{
calc_c_index(time_event, censored, var_name)
}
\arguments{
\item{time_event}{The time elapsed until an event occurs}

\item{censored}{Whether censoring occurred(0:censored, 1:event)}

\item{class}{Group separation criteria variable}
}
\value{
Variable-specific C-index for the event and 95\% confidence intervals
}
\description{
Calculate C-index for the specific variables
}
