% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_freq_rate.R
\name{calc_freq_rate}
\alias{calc_freq_rate}
\title{Calculate event frequency and rate per group}
\usage{
calc_freq_rate(time_event, time_vector, censored, class)
}
\arguments{
\item{time_event}{The time elapsed until an event occurs}

\item{time_vector}{A predefined time interval}

\item{censored}{Whether censoring occurred(0:censored, 1:event)}

\item{class}{Group separation criteria variable}
}
\value{
Frequency and proportion of events by group and the overall results of the Kaplan-Meier method
}
\description{
Calculate event frequency and rate per group
}
