% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hazard_ratio.R
\name{calc_hazard_ratio}
\alias{calc_hazard_ratio}
\title{Calculate the hazard ratio for the event by group}
\usage{
calc_hazard_ratio(time_event, censored, class)
}
\arguments{
\item{time_event}{The time elapsed until an event occurs}

\item{censored}{Whether censoring occurred(0:censored, 1:event)}

\item{class}{Group separation criteria variable}
}
\value{
Hazard ratio for the event by group
}
\description{
Calculate the hazard ratio for the event by group
}
