% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rmst.R
\name{calc_rmst}
\alias{calc_rmst}
\title{Calculate RMST for the event by group}
\usage{
calc_rmst(time_event, time_follow, censored, class)
}
\arguments{
\item{time_event}{The time elapsed until an event occurs}

\item{time_follow}{Follow-up period}

\item{censored}{Whether censoring occurred(0:censored, 1:event)}

\item{class}{Group separation criteria variable}
}
\value{
Group-specific RMST for the event and 95\% confidence intervals
}
\description{
Calculate RMST for the event by group
}
