% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_surv_rate.R
\name{calc_surv_rate}
\alias{calc_surv_rate}
\title{Calculate the survival rate for the event and test difference in survival rates by group}
\usage{
calc_surv_rate(time_event, time_follow, censored, class)
}
\arguments{
\item{time_event}{The time elapsed until an event occurs}

\item{time_follow}{Follow-up period}

\item{censored}{Whether censoring occurred(0:censored, 1:event)}

\item{class}{Group separation criteria variable}
}
\value{
Group-specific survival rates for the event and 95\% confidence intervals, the p-value of the log-rank test, and the overall results of the Kaplan-Meier method
}
\description{
Calculate the survival rate for the event and test difference in survival rates by group
}
