% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_surv_plot.R
\name{create_surv_plot}
\alias{create_surv_plot}
\title{Create Kaplane-Meier survival plot}
\usage{
create_surv_plot(km_res, filepath, title, x, y, legend_title, legend_labels)
}
\arguments{
\item{km_res}{Results of Kaplan-Meier method}

\item{filepath}{The file location where the survival curve will be saved}

\item{title}{The title of the survival curve}

\item{x}{The name of the x-axis}

\item{y}{The name of the y-axis}

\item{legend_title}{The title of the legend}

\item{legend_labels}{The names of legend labels}
}
\description{
Create Kaplane-Meier survival plot
}
