% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_data.R
\docType{data}
\name{ex_data}
\alias{ex_data}
\title{Generate Example Data}
\format{
A data frame with the following variables:
\describe{
\item{subjid}{The ID of the subject}
\item{censored}{Whether censoring occurred}
\item{time_event}{The time elapsed until an event occurs}
\item{age_class}{Group separation criteria variable}
}
}
\source{
Data created via simulation.
}
\usage{
ex_data
}
\description{
A simulated dataset created for use in the function examples.
}
\keyword{datasets}
