% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_surv_plot.R
\name{create_surv_plot}
\alias{create_surv_plot}
\title{Create Kaplane-Meier survival plot}
\usage{
create_surv_plot(km_res, filepath, title, x, y)
}
\arguments{
\item{time_event}{The time elapsed until an event occurs}

\item{censored}{Whether censoring occurred(0:censored, 1:event)}

\item{class}{Group separation criteria variable}
}
\description{
Create Kaplane-Meier survival plot
}
